add BatchingExecutor mixin trait to akka.dispatch, unused
For now this trait isn't used anywhere and the test is disabled;
just pushing it as something we can build on. It needs to be
hooked in to configuring the ExecutorService probably, and
we are also considering whether to make it a public API and
what that API would be like.