fix stupid bug in LARS, see #3467
When switching the enqueueing logic to use the actual time instead of
the current tick number (which had been the purpose of the whole
rewrite) I managed to make the same mistake I did half a year ago again:
enqueueing into the same slot that has just been processed needs to
subtract one round while enqueueing into the slot that is about to be
processed MUST NOT DO THIS.

There was another bug in there related to differences between tick
number and actual time and remaining ticks of the task. I hope that the
code is now clear and simple enough as to be actually correct. As usual,
all tests passing.